requirements.txt
absl-py @ file:///croot/absl-py_1714140470852/work
antlr4-python3-runtime @ file:///home/conda/feedstock_root/build_artifacts/antlr-python-runtime-meta_1638309185939/work
Bottleneck @ file:///croot/bottleneck_1731058641041/work
click @ file:///croot/click_1744271578095/work
Deprecated @ file:///work/ci_py311/deprecated_1676830461160/work
filelock @ file:///croot/filelock_1744281381737/work
fsspec @ file:///croot/fsspec_1747117721425/work
gmpy2 @ file:///croot/gmpy2_1738085463648/work
grpcio @ file:///croot/grpc-split_1742490263735/work
h5py @ file:///croot/h5py_1737051113867/work
hydra-core @ file:///home/conda/feedstock_root/build_artifacts/hydra-core_1736934708722/work
importlib_metadata @ file:///croot/importlib_metadata-suite_1732633488278/work
Jinja2 @ file:///croot/jinja2_1741710844255/work
lightning-utilities @ file:///croot/lightning-utilities_1734103441276/work
lit @ file:///croot/llvm-package_1741800901749/work/llvm/utils/lit
Markdown @ file:///croot/markdown_1746114451035/work
markdown-it-py @ file:///croot/markdown-it-py_1684279902645/work
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
mdurl @ file:///work/ci_py311/mdurl_1676827289376/work
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1737039604450/work
numexpr @ file:///croot/numexpr_1730215937391/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp311-cp311-linux_x86_64.whl#sha256=5f96f274d410a1682519282ae769c877d32fdbf171aa8badec7bf5e1d3a1748a
omegaconf @ file:///home/conda/feedstock_root/build_artifacts/omegaconf_1670575376789/work
opentelemetry-api @ file:///croot/opentelemetry-api_1743171649300/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///croot/pandas_1732735089971/work/dist/pandas-2.2.3-cp311-cp311-linux_x86_64.whl#sha256=884420a58b11b4c9b7f4df21046caaae9229522b672b557f0f98c84251141fc5
protobuf @ file:///croot/protobuf_1742419705443/work/bazel-bin/python/dist/protobuf-5.29.3-cp311-abi3-linux_x86_64.whl#sha256=8e447a0e89a4eb6aec01414ffb2b6e0032288e10ce1cf0a8129e73be2c0c34fb
Pygments @ file:///croot/pygments_1744664109463/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv @ file:///croot/python-dotenv_1745610241282/work
pytorch-lightning @ file:///croot/pytorch-lightning_1741103563308/work
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1737454647378/work
rich @ file:///croot/rich_1732638981168/work
rootutils @ file:///home/conda/feedstock_root/build_artifacts/rootutils_1744916563484/work
six @ file:///croot/six_1744271502820/work
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1745946051654/work
tensorboard @ file:///croot/tensorboard_1742506702034/work/tensorboard-2.19.0-py3-none-any.whl#sha256=5e71b98663a641a7ce8a6e70b0be8e1a4c0c45d48760b076383ac4755c35b9a0
tensorboard_data_server @ file:///croot/tensorboard-data-server_1724172556187/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl#sha256=c166a9f297fa12121b89b94be4a4cc150aa7996c9598410ba057ecd592eea8c7
tensorboardX @ file:///croot/tensorboardx_1722369114844/work
torch @ file:///opt/conda/conda-bld/libtorch_1745854776362/work
torchmetrics @ file:///croot/torchmetrics_1719910137958/work
tqdm @ file:///croot/tqdm_1738943501192/work
triton @ file:///croot/triton_1741878695212/work/python
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1744302253/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
Werkzeug @ file:///croot/werkzeug_1737448706492/work
wrapt @ file:///croot/wrapt_1736540904746/work
zipp @ file:///croot/zipp_1732630741423/work
